Rename History testing classes
